+++ src/StateSmith/Runner/DiServiceProvider.cs
@@ -44,7 +44,7 @@ public class DiServiceProvider : IDisposable
             services.AddSingleton<DiServiceProvider>(this); // todo_low remove. See https://github.com/StateSmith/StateSmith/issues/97
             services.AddSingleton<SmRunnerInternal>();
             services.AddSingleton<SmTransformer, StandardSmTransformer>();
-            services.AddSingleton<Expander>();
+            services.AddSingleton<IExpander, Expander>();
             services.AddSingleton<InputSmBuilder>();
             services.AddSingleton<IConsolePrinter, ConsolePrinter>();
             services.AddSingleton<ExceptionPrinter>();